requirements.txt
google-cloud-vision
requests
Pillow